Teach "export" command to output to stdout
Running an external script per item turns out to be quite slow.  On
a test project printing output to stdout seems to be more than 6 times
faster and for vifm project it's 16 (!) times faster.

Also script seems to be way easier to write as there is only one process
and it gets all the data.