Reduce limit for spawning pager by one row
Shell prompt is usually one line, so `ROWS - 1` will cause pager to be
spawned when at least one line of output would be hidden.

We don't have any means for accounting for two or more line prompts,
although there could be a confirmation for this (like number of reserved
lines with `1` as a default.